refactor: Folder Structure to account for javafx
