+++ src/main/java/oak/Launcher.java
@@ -1,6 +1,7 @@
 package oak;
 
 import javafx.application.Application;
+import oak.javafx.Main;
 
 public class Launcher {
     public static void main(String[] args) {
