+++ src/main/java/oak/controller/OakController.java
@@ -1,0 +1,7 @@
+package oak.controller;
+
+public class OakController {
+    public String getResponse(String input) {
+        return "Duke heard: " + input;
+    }
+}
